Switched to real market API. Replaced int with long

+++ app/src/main/java/alex/com/bitcoinmanager/api/service/response/OrderBookServiceResponse.java
@@ -5,7 +5,7 @@ package alex.com.bitcoinmanager.api.service.response;
  */
 
 public class OrderBookServiceResponse {
-    public int sequence;
+    public long sequence;
     public Object[][] bids;
     public Object[][] asks;
 }
